Flatten service-worker-mock
