+++ packages/service-worker-mock/models/NotificationEvent.js
