+++ packages/service-worker-mock/models/ServiceWorkerRegistration.js
